requirements.txt
absl-py==1.3.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
bcj-cffi==0.5.1
beautifulsoup4==4.11.1
Bottleneck==1.3.5
Brotli==1.0.9
brotlicffi==1.0.9.2
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
conllu==4.4.1
cryptography==38.0.1
datasets==2.6.1
dill==0.3.4
et-xmlfile==1.1.0
filelock==3.8.0
flatbuffers==1.12
frozenlist==1.2.0
fsspec==2022.7.1
gast==0.4.0
gdown==4.5.3
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-trans-new==1.1.9
googletrans==4.0.0rc1
grpcio==1.50.0
h11==0.9.0
h2==3.2.0
h5py==3.7.0
hpack==3.0.0
hstspreload==2022.11.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.10.1
hyperframe==5.2.0
idna==2.10
importlib-metadata==5.0.0
joblib==1.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
multiprocess==0.70.12.2
multivolumefile==0.2.3
nlpaug==1.1.11
numexpr==2.8.3
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pip==22.3
protobuf==3.19.6
py7zr==0.16.1
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.15.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppmd==0.16.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
pyzstd==0.14.4
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
setuptools==65.5.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
texttable==1.6.4
threadpoolctl==3.1.0
tokenizers==0.13.1
torch==1.13.0
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
xxhash==2.0.2
yarl==1.8.1
zipp==3.10.0
